requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scipy>=1.7.0
scikit-learn>=1.0.0
tqdm>=4.62.0
opencv-python>=4.5.0
dask>=2021.9.0
numba>=0.56.0
streamlit>=1.0.0
plotly>=5.0.0
xarray>=0.20.0
netCDF4>=1.6.0
h5py>=3.0.0
pyhdf>=0.11.0
pyyaml>=5.4.0
schedule>=1.1.0